STYLE: Fixed Warninings

+++ smtk/attribute/Definition.cxx
@@ -246,7 +246,7 @@ smtk::attribute::DefinitionPtr Definition::pointer() const
 smtk::attribute::ItemDefinitionPtr Definition::itemDefinition(int ith) const
 {
   // Is the item in this defintion?
-  if (ith >= this->m_baseItemOffset)
+  if (ith >= (int)this->m_baseItemOffset)
     {
     return this->m_itemDefs[static_cast<std::size_t>(ith-this->m_baseItemOffset)];
     }
